Add option to search for all biomes
